make it typescript
